+++ projects/angular2-smart-table/src/lib/lib/settings.ts
@@ -26,6 +26,7 @@ export interface Settings {
 
 export interface Filter {
   inputClass?: string;
+  debounceTime?: number;
 }
 
 export interface Expand {