+++ projects/demo/src/app/pages/examples/custom-edit-view/custom-filter.component.ts
@@ -27,7 +27,7 @@ export class CustomFilterComponent extends DefaultFilter implements OnInit, OnCh
     this.inputControl.valueChanges
       .pipe(
         distinctUntilChanged(),
-        debounceTime(this.delay),
+        debounceTime(this.debounceTime),
       )
       .subscribe(value => {
         this.query = value;